requirements.txt
azure-cognitiveservices-speech==1.38.0
azure-storage-blob==12.20.0
fastapi==0.115.0
httpx==0.27.0
numpy==1.26.4
openai==1.35.10
pillow==10.4.0
pydantic==2.10.4
pydub==0.25.1
python-dotenv==1.0.1
python-json-logger==2.0.7
requests==2.32.3
starlette==0.38.6
streamlit==1.38.0
uvicorn[standard]==0.30.6
wordcloud==1.9.3
